nav: Use 16-bit data instead of 32 bit floats to reduce portal cost field data size

+++ src/navigation/nav_private.h
@@ -54,7 +54,7 @@ struct nav_private{
     /* Owns the backing storage for portal_travel_costs (live map only; NULL in
      * snapshots, which share the live map's chunk pointers read-only).
      */
-    float                (*portal_travel_costs_mem)[FIELD_RES_R][FIELD_RES_C];
+    uint16_t             (*portal_travel_costs_mem)[FIELD_RES_R][FIELD_RES_C];
     /* Private cache for fields and other computation-heavy intermediate data */
     struct fieldcache_ctx *fieldcache;
     /* Data used for fieldcache invalidation */
